Refactor guests and sessions handling in routes and models
The guest route handlers related to session sign-in and sign-out have been removed and transferred to a new sessions route. New controller and model files have been added specifically for sessions. The logic that was previously embedded in the guests model for managing session details is now in the sessions model.